+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## Unreleased
 
 - Minor: Parse slayer kill log for more accurate kill count reporting. (#718)
+
+## 1.11.8
+
 - Minor: Add Hueycoatl pet data to PetNotifier. (#719)
+- Minor: Update Kill Count Notifier to respect Yama "success" message. (#728)
+- Minor: Include Yami pet name in pet notifications. (#730)
 
 ## 1.11.7